+++ bootstrap4/bootstrap.py
@@ -12,13 +12,15 @@ from django.conf import settings
 DBS3_SET_REQUIRED_SET_DISABLED = DJANGO_VERSION[0] < 2 and DJANGO_VERSION[1] < 10
 
 # Default settings
-BOOTSTRAP3_DEFAULTS = {
-    'jquery_url': '//code.jquery.com/jquery.min.js',
-    'base_url': '//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/',
+BOOTSTRAP4_DEFAULTS = {
+    'jquery_url': '//code.jquery.com/jquery-3.1.1.slim.min.js',
+    'jquery_integrity': 'sha384-A7FZj7v+d/sdmMqp/nOQwliLvUsJfDHW+k9Omg/a/EheAdgtzNs3hpfag6Ed950n',
+    'base_url': '//maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/',
     'css_url': None,
+    'css_integrity': 'sha384-rwoIResjU2yc3z8GV/NPeZWAv56rSmLldC3R/AZzGRnGxQQKnKkoFVhFQhNUwEyJ',
     'theme_url': None,
     'javascript_url': None,
-    'javascript_in_head': False,
+    'javascript_integrity': "sha384-vBWWzlZJ8ea9aCX4pEW3rVHjgjt7zpkNpZk+02D9phzyeVkE+jo0ieGizqPLForn",
     'include_jquery': False,
     'horizontal_label_class': 'col-md-3',
     'horizontal_field_class': 'col-md-9',
@@ -28,35 +30,35 @@ BOOTSTRAP3_DEFAULTS = {
     'error_css_class': 'has-error',
     'success_css_class': 'has-success',
     'formset_renderers': {
-        'default': 'bootstrap3.renderers.FormsetRenderer',
+        'default': 'bootstrap4.renderers.FormsetRenderer',
     },
     'form_renderers': {
-        'default': 'bootstrap3.renderers.FormRenderer',
+        'default': 'bootstrap4.renderers.FormRenderer',
     },
     'field_renderers': {
-        'default': 'bootstrap3.renderers.FieldRenderer',
-        'inline': 'bootstrap3.renderers.InlineFieldRenderer',
+        'default': 'bootstrap4.renderers.FieldRenderer',
+        'inline': 'bootstrap4.renderers.InlineFieldRenderer',
     },
 }
 
 if DBS3_SET_REQUIRED_SET_DISABLED:
-    BOOTSTRAP3_DEFAULTS.update({
+    BOOTSTRAP4_DEFAULTS.update({
         'set_required': True,
         'set_disabled': False,
     })
 
 # Start with a copy of default settings
-BOOTSTRAP3 = BOOTSTRAP3_DEFAULTS.copy()
+BOOTSTRAP4 = BOOTSTRAP4_DEFAULTS.copy()
 
 # Override with user settings from settings.py
-BOOTSTRAP3.update(getattr(settings, 'BOOTSTRAP3', {}))
+BOOTSTRAP4.update(getattr(settings, 'BOOTSTRAP4', {}))
 
 
 def get_bootstrap_setting(setting, default=None):
     """
     Read a setting
     """
-    return BOOTSTRAP3.get(setting, default)
+    return BOOTSTRAP4.get(setting, default)
 
 
 def bootstrap_url(postfix):
@@ -72,6 +74,12 @@ def jquery_url():
     """
     return get_bootstrap_setting('jquery_url')
 
+def jquery_integrity():
+    """
+    Return the full url to jQuery file to use
+    """
+    return get_bootstrap_setting('jquery_integrity')
+
 
 def javascript_url():
     """
@@ -81,12 +89,26 @@ def javascript_url():
            bootstrap_url('js/bootstrap.min.js')
 
 
+def javascript_url_integrity():
+    """
+    Return the full url integrity to the Bootstrap JavaScript file
+    """
+    return get_bootstrap_setting('javascript_integrity') or \
+           BOOTSTRAP4_DEFAULTS['javascript_integrity']
+
 def css_url():
     """
     Return the full url to the Bootstrap CSS file
     """
     return get_bootstrap_setting('css_url') or \
            bootstrap_url('css/bootstrap.min.css')
+
+def css_url_integrity():
+    """
+    Return the full css integrity to the Bootstrap JavaScript file
+    """
+    return get_bootstrap_setting('css_integrity') or \
+           BOOTSTRAP4_DEFAULTS['css_integrity']
 
 
 def theme_url():