+++ demo/demo/settings.py
@@ -7,9 +7,9 @@ import sys
 BASE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 
 # Include BOOTSTRAP3_FOLDER in path
-BOOTSTRAP3_FOLDER = os.path.abspath(os.path.join(BASE_DIR, '..', 'bootstrap3'))
-if BOOTSTRAP3_FOLDER not in sys.path:
-    sys.path.insert(0, BOOTSTRAP3_FOLDER)
+BOOTSTRAP4_FOLDER = os.path.abspath(os.path.join(BASE_DIR, '..', 'bootstrap4'))
+if BOOTSTRAP4_FOLDER not in sys.path:
+    sys.path.insert(0, BOOTSTRAP4_FOLDER)
 
 DEBUG = True
 
@@ -19,7 +19,7 @@ DATABASES = {}
 
 # Hosts/domain names that are valid for this site; required if DEBUG is False
 # See https://docs.djangoproject.com/en/1.5/ref/settings/#allowed-hosts
-ALLOWED_HOSTS = ['localhost', '127.0.0.1', ]
+ALLOWED_HOSTS = ['localhost', '127.0.0.1','0.0.0.0' ]
 
 # Local time zone for this installation. Choices can be found here:
 # http://en.wikipedia.org/wiki/List_of_tz_zones_by_name
@@ -117,7 +117,7 @@ INSTALLED_APPS = (
     'django.contrib.messages',
     'django.contrib.staticfiles',
 
-    'bootstrap3',
+    'bootstrap4',
     'demo',
 )
 
@@ -151,9 +151,8 @@ LOGGING = {
 }
 
 # Settings for django-bootstrap3
-BOOTSTRAP3 = {
+BOOTSTRAP4 = {
     'set_required': False,  # For Django <= 1.8 only
-    'error_css_class': 'bootstrap3-error',
-    'required_css_class': 'bootstrap3-required',
-    'javascript_in_head': True,
+    'error_css_class': 'bootstrap4-error',
+    'required_css_class': 'bootstrap4-required',
 }